refactor(preset-mini,preset-wind): split and prioritize `x-none`, `x-normal` (#278)
