+++ packages/core/src/types.ts
@@ -13,6 +13,30 @@ export type RequiredByKey<T, K extends keyof T = keyof T> = FlatObjectTuple<Requ
 export type CSSObject = Record<string, string | number | undefined>
 export type CSSEntries = [string, string | number | undefined][]
 
+export type RGBAColorValue = [number, number, number, number] | [number, number, number]
+export type ParsedColorValue = {
+  /**
+   * Parsed color value.
+   */
+  color?: string
+  /**
+   * Parsed opacity value.
+   */
+  opacity: string
+  /**
+   * Color name.
+   */
+  name: string
+  /**
+   * Color scale. Preferrably 000 - 999
+   */
+  no: string
+  /**
+   * Color scale. Preferrably 000 - 999
+   */
+  rgba?: RGBAColorValue
+}
+
 export interface RuleContext<Theme extends {} = {}> {
   /**
    * Unprocessed selector from user input.
