+++ packages/preset-mini/src/rules/decoration.ts
@@ -1,21 +1,18 @@
-import type { Rule } from '@unocss/core'
-import { handler as h } from '../utils'
-import { colorResolver } from './color'
+import type { CSSObject, Rule } from '@unocss/core'
+import { colorResolver, handler as h } from '../utils'
 
 export const textDecorations: Rule[] = [
   ['underline', { 'text-decoration': 'underline' }],
   ['line-through', { 'text-decoration': 'line-through' }],
-  ['no-underline', { 'text-decoration': 'none' }],
   ['decoration-underline', { 'text-decoration': 'underline' }],
   ['decoration-line-through', { 'text-decoration': 'line-through' }],
-  ['decoration-none', { 'text-decoration': 'none' }],
 
   // style
   [/^(?:underline|decoration)-(solid|double|dotted|dashed|wavy)$/, ([, d]) => ({ 'text-decoration-style': d })],
 
   // thickness
   [/^(?:underline|decoration)-([^-]+)$/, ([, s]) => ({ 'text-decoration-thickness': ['auto', 'from-font'].includes(s) ? s : h.bracket.px(s) })],
-  [/^decoration-(.*)$/, ([, d]) => ({ 'text-decoration-thickness': h.bracket.px(d) })],
+  [/^decoration-(.+)$/, ([, d]) => ({ 'text-decoration-thickness': h.bracket.px(d) })],
 
   // offset
   [/^underline-offset-([^-]+)$/, ([, s]) => {
@@ -26,7 +23,7 @@ export const textDecorations: Rule[] = [
 
   // colors
   [/^(?:underline|decoration)-(.+)$/, (match, ctx) => {
-    const result = colorResolver('text-decoration-color', 'line')(match, ctx)
+    const result = colorResolver('text-decoration-color', 'line')(match, ctx) as CSSObject | undefined
     if (result) {
       return {
         '-webkit-text-decoration-color': result['text-decoration-color'],
@@ -34,5 +31,8 @@ export const textDecorations: Rule[] = [
       }
     }
   }],
-  [/^(?:underline|decoration)-op(?:acity)?-?(.+)$/m, ([, opacity]) => ({ '--un-line-opacity': h.bracket.percent(opacity) })],
+  [/^(?:underline|decoration)-op(?:acity)?-?(.+)$/, ([, opacity]) => ({ '--un-line-opacity': h.bracket.percent(opacity) })],
+
+  ['no-underline', { 'text-decoration': 'none' }],
+  ['decoration-none', { 'text-decoration': 'none' }],
 ]
