+++ packages/preset-mini/src/rules/grid.ts
@@ -3,7 +3,7 @@ import { toArray } from '@unocss/core'
 import type { Theme } from '../theme'
 import { handler as h } from '../utils'
 
-const calSize = (s: string, theme: Theme) => toArray(theme.fontSize?.[s] || h.bracket.rem(s))[0]
+const calSize = (s: string, theme: Theme) => toArray(theme.fontSize?.[s] || h.bracket.auto.rem(s))[0]
 
 const autoDirection = (selector: string, theme: Theme) => {
   if (selector === 'min')
@@ -19,10 +19,10 @@ export const grids: Rule[] = [
   // base
   ['grid', { display: 'grid' }],
   ['inline-grid', { display: 'inline-grid' }],
-  [/^(?:grid-)?col-start-([\w.-]+)$/, ([, v]) => ({ 'grid-column-start': `${v}` })],
-  [/^(?:grid-)?col-end-([\w.]+)$/, ([, v]) => ({ 'grid-column-end': `${v}` })],
-  [/^(?:grid-)?row-start-([\w.-]+)$/, ([, v]) => ({ 'grid-row-start': `${v}` })],
-  [/^(?:grid-)?row-end-([\w.-]+)$/, ([, v]) => ({ 'grid-row-end': `${v}` })],
+  [/^(?:grid-)?col-start-([\w.-]+)$/, ([, v]) => ({ 'grid-column-start': v })],
+  [/^(?:grid-)?col-end-([\w.]+)$/, ([, v]) => ({ 'grid-column-end': v })],
+  [/^(?:grid-)?row-start-([\w.-]+)$/, ([, v]) => ({ 'grid-row-start': v })],
+  [/^(?:grid-)?row-end-([\w.-]+)$/, ([, v]) => ({ 'grid-row-end': v })],
   [/^(?:grid-)?(row|col)-(.+)$/, ([, d, v]) => {
     const key = d === 'row' ? 'grid-row' : 'grid-column'
     let raw = h.bracket(v)
@@ -42,11 +42,11 @@ export const grids: Rule[] = [
   }],
 
   // templates
-  [/^(?:grid-)?auto-cols-([\w.-]+)$/, ([, v], { theme }) => ({ 'grid-auto-columns': `${autoDirection(v, theme)}` })],
-  [/^(?:grid-)?auto-flow-([\w.-]+)$/, ([, v]) => ({ 'grid-auto-flow': `${v.replace('col', 'column').split('-').join(' ')}` })],
-  [/^(?:grid-)?auto-rows-([\w.-]+)$/, ([, v], { theme }) => ({ 'grid-auto-rows': `${autoDirection(v, theme)}` })],
-  [/^grid-cols-minmax-([\w.-]+)$/, ([, d]) => ({ 'grid-template-columns': `repeat(auto-fill, minmax(${d}, 1fr))` })],
-  [/^grid-rows-minmax-([\w.-]+)$/, ([, d]) => ({ 'grid-template-rows': `repeat(auto-fill, minmax(${d}, 1fr))` })],
+  [/^(?:grid-)?auto-cols-([\w.-]+)$/, ([, v], { theme }) => ({ 'grid-auto-columns': autoDirection(v, theme) })],
+  [/^(?:grid-)?auto-flow-([\w.-]+)$/, ([, v]) => ({ 'grid-auto-flow': v.replace('col', 'column').split('-').join(' ') })],
+  [/^(?:grid-)?auto-rows-([\w.-]+)$/, ([, v], { theme }) => ({ 'grid-auto-rows': autoDirection(v, theme) })],
+  [/^grid-cols-minmax-([\w.-]+)$/, ([, d]) => ({ 'grid-template-columns': `repeat(auto-fill,minmax(${d},1fr))` })],
+  [/^grid-rows-minmax-([\w.-]+)$/, ([, d]) => ({ 'grid-template-rows': `repeat(auto-fill,minmax(${d},1fr))` })],
   [/^grid-cols-(\d+)$/, ([, d]) => ({ 'grid-template-columns': `repeat(${d},minmax(0,1fr))` })],
   [/^grid-rows-(\d+)$/, ([, d]) => ({ 'grid-template-rows': `repeat(${d},minmax(0,1fr))` })],
   [/^grid-cols-\[(.+)\]$/, ([, v]) => ({ 'grid-template-columns': v.replace(/,/g, ' ') })],
