+++ packages/preset-wind/src/rules/spacing.ts
@@ -1,11 +1,5 @@
-import type { CSSEntries, Rule } from '@unocss/core'
-import { directionMap, handler as h } from '@unocss/preset-mini/utils'
-
-const directionSize = (prefix: string) => ([_, direction, size]: string[]): CSSEntries | undefined => {
-  const v = h.bracket.rem.fraction.cssvar(size)
-  if (v)
-    return directionMap[direction].map(i => [prefix + i, v])
-}
+import type { Rule } from '@unocss/core'
+import { directionSize } from '@unocss/preset-mini/utils'
 
 export const spaces: Rule[] = [
   [/^space-?([xy])-?(-?.+)$/, (match) => {
