+++ test/preset-mini-targets.ts
@@ -1,27 +1,46 @@
 export const presetMiniTargets: string[] = [
+  // align
+  'vertical-baseline',
+  'align-text-bottom',
+  'v-top',
+  'v-mid',
+  'text-left',
+
+  // behaviors
+  'outline-none',
+  'outline',
+  'outline-gray',
+  'outline-gray-400',
+  'outline-size-4',
+  'outline-width-4',
+  'outline-offset-4',
+  'outline-unset',
+  'outline-solid',
+  'outline-color-red-1',
+  'outline-width-10px',
+  'outline-inset',
+  'outline-110',
+  'outline-blue-2',
+  'outline-none',
+  'outline-[var(--red)]',
+  'outline-size-[var(--width)]',
+  'outline-offset-[var(--offset)]',
+  'appearance-none',
+  'placeholder-color-red-1',
+  'placeholder-transparent',
+  'placeholder-opacity-60',
+  'will-change-transform',
+  'will-change-unset',
+
   // border
-  'border-[#124]',
-  'border-[2em]',
-  'border-[calc(1em-1px)]',
+  'b-2',
+  'border',
+  'border-double',
+  'border-none',
   'border-4',
-  'border-black/10',
-  'border-blue',
-  'border-red-100',
-  'border-red-200/10',
-  'border-red-300/20',
-  'border-red100',
-  'border-red2',
-  'border-[var(--color)]',
   'border-b',
   'border-x',
-  'border-green-100/20',
   'border-t-2',
-  'border-y-red',
-  'border-x-[rgb(1,2,3)]/[0.5]',
-  'border',
-  'border-$color',
-  'border-t-$color',
-  'border-x-$color',
   'rounded-[4px]',
   'rounded-1/2',
   'rounded-full',
@@ -30,6 +49,172 @@ export const presetMiniTargets: string[] = [
   'rounded-t-sm',
   'rounded-tr',
   'rounded',
+  'border-rounded',
+
+  // color, op
+  'op-10',
+  'opacity-0',
+  'opacity-$opa',
+  'color-blue',
+  'color-blue-400',
+  'color-blue-400/10',
+  'color-blue/10',
+  'color-blue-gray',
+  'color-blue-gray-400',
+  'color-bluegray',
+  'color-bluegray-400',
+  'color-blue-gray-400/10',
+  'color-blue-gray/10',
+  'color-bluegray-400/10',
+  'color-bluegray/10',
+  'text-black/10',
+  'text-rose',
+  'text-red-100',
+  'text-red-200/10',
+  'text-red-300/20',
+  'text-red100',
+  'text-red2',
+  'text-opacity-[13.3333333%]',
+  'text-[var(--color)]',
+  'text-[#124]',
+  'text-[2em]',
+  'text-[calc(1em-1px)]',
+
+  // color - bg
+  'bg-[#153]/10',
+  'bg-[#1533]',
+  'bg-[#1533]/10',
+  'bg-[rgba(1,2,3,0.5)]',
+  'bg-#452233/40',
+  'bg-red-100',
+  'bg-teal-100/55',
+  'bg-teal-200:55',
+  'bg-teal-300:[.55]',
+  'bg-teal-400/[.55]',
+  'bg-teal-500/[55%]',
+  'bg-hex-452233/40',
+  'bg-opacity-45',
+
+  // color - border
+  'border-[#124]',
+  'border-[2em]',
+  'border-[calc(1em-1px)]',
+  'border-black/10',
+  'border-blue',
+  'border-red-100',
+  'border-red-200/10',
+  'border-red-300/20',
+  'border-red100',
+  'border-red2',
+  'border-[var(--color)]',
+  'border-green-100/20',
+  'border-y-red',
+  'border-x-[rgb(1,2,3)]/[0.5]',
+
+  // color - ring
+  'ring-red2',
+  'ring-red2/5',
+  'ring-op-5',
+  'ring-offset-red2',
+  'ring-offset-red2/5',
+  'ring-offset-op-5',
+
+  // decoration
+  'decoration-none',
+  'decoration-transparent',
+  'decoration-purple/50',
+  'decoration-5',
+  'underline',
+  'underline-dashed',
+  'underline-red-500',
+  'underline-op80',
+  'underline-auto',
+  'underline-5',
+  'underline-1rem',
+  'underline-offset-auto',
+  'underline-offset-4',
+  'underline-offset-0.6rem',
+
+  // flex, gap
+  'flex-[hi]',
+  'flex',
+  'flex-row',
+  'flex-col-reverse',
+  'gap-4',
+  'gap-x-1',
+  'flex-gap-y-1',
+  'grid-gap-y-1',
+
+  // grid
+  'grid-cols-$1',
+  'grid-cols-[1fr,2fr,100px,min-content]',
+  'grid-cols-2',
+  'grid-rows-[1fr,2fr,100px,min-content]',
+  'grid-rows-3',
+  'grid',
+  'auto-rows-min',
+  'auto-rows-fr',
+  'row-auto',
+  'row-span-[hi]',
+  'row-[span_1/span_2]',
+  'auto-cols-auto',
+  'auto-rows-auto',
+  'col-span-1',
+  'row-span-full',
+  'row-end-1',
+  'row-start-full',
+  'auto-flow-cols-dense',
+  'grid-cols-minmax-1rem',
+  'grid-rows-minmax-100px',
+
+  // layout
+  'of-y-visible',
+  'overflow-auto',
+  'overflow-x-scroll',
+
+  // position
+  'justify-start',
+  'justify-center',
+  'justify-items-end',
+  'justify-self-stretch',
+  'order-first',
+  'order-none',
+  'content-start',
+  'content-center',
+  'items-end',
+  'self-stretch',
+  'place-content-stretch',
+  'place-items-stretch',
+  'place-self-stretch',
+
+  // position - insets
+  'top-0',
+  'top-$top-height',
+  'inset-x-5',
+  'float-left',
+  'float-none',
+  'clear-both',
+  'clear-none',
+  'z-0',
+  'z-1',
+  'z-100',
+  'z-auto',
+  'box-content',
+  'box-border',
+
+  // ring, shadow
+  'ring',
+  'ring-10',
+  'ring-offset',
+  'ring-offset-4',
+  'ring-offset-green5',
+  'ring-inset',
+  'shadow',
+  'shadow-transparent',
+  'shadow-current',
+  'shadow-none',
+  'shadow-xl',
+  'shadow-green-500',
 
   // size
   'h-auto',
@@ -60,33 +245,65 @@ export const presetMiniTargets: string[] = [
   'h-[calc(1000px-4rem)]',
   'w-[calc(calc(100px*10)-4rem)]',
 
-  // background
-  'bg-[#153]/10',
-  'bg-[#1533]',
-  'bg-[#1533]/10',
-  'bg-[rgba(1,2,3,0.5)]',
-  'bg-#452233/40',
-  'bg-red-100',
-  'bg-teal-100/55',
-  'bg-teal-200:55',
-  'bg-teal-300:[.55]',
-  'bg-teal-400/[.55]',
-  'bg-teal-500/[55%]',
+  // size - ar
+  'aspect-ratio-auto',
+  'aspect-ratio-3/2',
+  'aspect-ratio-0.7',
+  'aspect-ratio-$var',
+  'aspect-ratio-[auto_16/9]',
 
-  // decoration
-  'decoration-none',
-  'decoration-transparent',
-  'decoration-purple/50',
-  'underline',
-  'underline-dashed',
-  'underline-red-500',
-  'underline-op80',
-  'underline-auto',
-  'underline-5',
-  'underline-1rem',
-  'underline-offset-auto',
-  'underline-offset-4',
-  'underline-offset-0.6rem',
+  // spacing
+  'p-2',
+  'p-t-2',
+  'p2',
+  'pl-10px',
+  'pt-2',
+  'pt2',
+  'pt-$title2',
+  'm-[3em]',
+  'm-0',
+  'm-1/2',
+  'm-auto',
+  'mt-[-10.2%]',
+  'mt-$height',
+  'my-auto',
+
+  // static
+  'contents',
+  'backface-hidden',
+  'cursor-pointer',
+  'pointer-events-none',
+  'pointer-events-auto',
+  'resize',
+  'resize-none',
+  'select-all',
+  'select-none',
+  'whitespace-pre-wrap',
+  'ws-nowrap',
+  'content-empty',
+  'break-normal',
+  'text-clip',
+  'case-upper', // !
+  'case-normal', // !
+  'italic',
+  'antialiased',
+
+  // svg
+  'fill-none',
+  'fill-current',
+  'fill-green-400',
+  'fill-opacity-80',
+  'fill-[#123]',
+  'fill-[1rem]',
+  'stroke-none',
+  'stroke-current',
+  'stroke-green-400',
+  'stroke-opacity-80',
+  'stroke-[#123]',
+  'stroke-[1rem]',
+  'stroke-size-1',
+  'stroke-size-1px',
+  'stroke-size-[1rem]',
 
   // transforms
   'transform',
@@ -94,43 +311,14 @@ export const presetMiniTargets: string[] = [
   'transform-cpu',
   'transform-none',
   'translate-y-1/4',
-  '-translate-y-1/2',
   'translate-y-px',
   'translate-full',
-  '-translate-full',
   'translate-x-full',
-  '-translate-x-full',
-  'hover:translate-x-3',
-  'peer-checked:translate-x-[var(--reveal)]',
-  'before:translate-y-full',
   'rotate-1deg',
-  '-rotate-2',
   'rotate-[var(--spin)]',
-  'active:scale-4',
   'preserve-3d',
   'preserve-flat',
-
-  // grid
-  'grid-cols-$1',
-  'grid-cols-[1fr,2fr,100px,min-content]',
-  'grid-cols-2',
-  'grid-rows-[1fr,2fr,100px,min-content]',
-  'grid-rows-3',
-  'grid',
-  'auto-rows-min',
-  'auto-rows-fr',
-  'row-auto',
-  'row-span-[hi]',
-  'row-[span_1/span_2]',
-  'auto-cols-auto',
-  'auto-rows-auto',
-  'col-span-1',
-  'row-span-full',
-  'row-end-1',
-  'row-start-full',
-  'auto-flow-cols-dense',
-  'grid-cols-minmax-1rem',
-  'grid-rows-minmax-100px',
+  'origin-top-left',
 
   // transition
   'transition-none',
@@ -140,7 +328,99 @@ export const presetMiniTargets: string[] = [
   'transition-property-all',
   'transition-200',
   'transition',
+  'property-all',
+  'duration-111',
+  'ease-out',
 
-  // transition-timing-function
+  // transition - timings
   'ease-linear',
+
+  // typography
+  'font-mono',
+  'text-4xl',
+  'text-base',
+  'text-lg',
+  'text-size-[2em]',
+  'font-thin',
+  'fw-900',
+  'leading-2',
+  'lh-[1.5]',
+  'tracking-wide',
+  'tracking-[2/5]',
+  'word-spacing-1',
+  'word-spacing-wide',
+  'word-spacing-2',
+  'tab',
+  'tab-6',
+  'tab-inherit',
+  'indent',
+  'indent-1/2',
+  'indent-lg',
+  'text-stroke-6',
+  'text-stroke-sm',
+  'text-stroke-blue-500',
+  'text-stroke-op-60',
+  'text-shadow',
+  'text-shadow-lg',
+  'text-shadow-none',
+  'text-shadow-$var',
+
+  // variables
+  'bg-$test-variable',
+  'border-$color',
+  'border-t-$color',
+  'border-x-$color',
+  'color-$red',
+  'items-$size',
+  'object-$fit',
+  'tab-$tabprop',
+
+  // dark/light
+  'dark:not-odd:text-red',
+  'dark:text-xl',
+  'light:text-sm',
+
+  // variants
+  '-rotate-2',
+  '-translate-full',
+  '-translate-x-full',
+  '-translate-y-1/2',
+  'active:scale-4',
+  'before:translate-y-full',
+  'hover:translate-x-3',
+  'peer-checked:translate-x-[var(--reveal)]',
+
+  'before:translate-y-full',
+  '-rotate-2',
+  'active:scale-4',
+  '!hover:px-10',
+  '!m-$c-m',
+  '!p-5px',
+  '-gap-y-5',
+  '-m-auto',
+  '-mb-px',
+  '-p-px',
+  '-z-1',
+  'all:m-auto',
+  'at-2xl:m2',
+  'at-lg:m2',
+  'at-sm:m1',
+  'children:m-auto',
+  'disabled:op50',
+  'first:p-2',
+  'group-focus:p-4',
+  'group-hover:group-focus:text-center',
+  'hover:!p-1',
+  'hover:not-first:checked:bg-red/10',
+  'hover:p-5',
+  'lt-lg:m2',
+  'lt-sm:m1',
+  'md:!hidden',
+  'md:m-1',
+  'next:mt-0',
+  'not-hover:p-3',
+  'peer-checked:bg-blue-500',
+  'peer-not-placeholder-shown:text-2xl',
+  'sm:m-1',
+  'sm:m1',
 ]
