+++ test/preset-wind-targets.ts
@@ -1,77 +1,8 @@
-
 export const presetWindiTargets: string[] = [
-  // tables
-  'border-collapse',
-  'border-separate',
-  // object-position
-  'object-center',
-  'object-ct',
-  'object-cb',
-  // divide
-  'divide',
-  'divide-y-4',
-  'divide-x-4',
-  'divide-x-reverse',
-  'divide-green-500',
-  'divide-opacity-50',
-  'divide-dashed',
-  'divide-dotted',
-  'divide-transparent',
-  'divide-current',
-  // backdrop filters
-  'backdrop-filter',
-  'backdrop-blur',
-  'backdrop-blur-md',
-  'backdrop-blur-4',
-  'backdrop-brightness-0',
-  'backdrop-brightness-60',
-  'backdrop-contrast-125',
-  'backdrop-grayscale',
-  'backdrop-grayscale-90',
-  'backdrop-hue-rotate-0',
-  'backdrop-hue-rotate-360',
-  '-backdrop-hue-rotate-90',
-  'backdrop-invert',
-  'backdrop-invert-90',
-  'backdrop-saturate',
-  'backdrop-saturate-30',
-  'backdrop-sepia',
-  'backdrop-sepia-80',
-  'line-clamp-7',
-  'line-clamp-100',
-  'line-clamp-none',
-  // isolation
-  'isolate',
-  'isolate-auto',
-  // filters
-  'filter',
-  'blur',
-  'blur-md',
-  'blur-4',
-  'brightness-0',
-  'brightness-60',
-  'contrast-125',
-  'drop-shadow',
-  'drop-shadow-[0_4px_3px_#000]',
-  'drop-shadow-none',
-  'drop-shadow-md',
-  'grayscale',
-  'grayscale-90',
-  'hue-rotate-0',
-  'hue-rotate-360',
-  '-hue-rotate-90',
-  'invert',
-  'invert-90',
-  'saturate',
-  'saturate-30',
-  'sepia',
-  'sepia-80',
   // animation
   'animate-none',
-  '!animate-ping',
   'animate-pulse',
   'animate-pulse-alt',
-  'hover:animate-bounce',
   'animate-300',
   'animate-100s',
   'animate-duration-100',
@@ -88,4 +19,211 @@ export const presetWindiTargets: string[] = [
   'animate-count-2.4',
   'animate-iteration-count-2',
   'animate-iteration-count-2-4-infinity',
+
+  // background
+  'bg-auto',
+  'bg-blend-normal',
+  'bg-blend-color-burn',
+  'bg-blend-luminosity',
+  'bg-bottom',
+  'bg-clip-border',
+  'bg-clip-text',
+  'bg-cover',
+  'bg-fixed',
+  'bg-gradient-to-t',
+  'bg-gradient-to-tl',
+  'bg-local',
+  'bg-no-repeat',
+  'bg-none',
+  'bg-origin-border',
+  'bg-repeat-space',
+  'bg-right-bottom',
+  'bg-scroll',
+
+  // bg gradient
+  'from-current',
+  'from-green-500',
+  'from-green-500/50',
+  'from-transparent',
+  'to-current',
+  'to-green-500',
+  'to-green-500/50',
+  'to-transparent',
+  'via-current',
+  'via-green-500',
+  'via-green-500/30',
+  'via-transparent',
+  'via-opacity-30',
+
+  // behaviors
+  'list-none',
+  'list-disc',
+  'list-outside',
+  'decoration-slice',
+  'accent-op-90',
+  'accent-red',
+  'caret-op-90',
+  'caret-red',
+  'image-render-pixel',
+  'overscroll-x-auto',
+  'overscroll-contain',
+  'overscroll-none',
+  'scroll-auto',
+
+  // columns
+  'columns-3',
+  'columns-5em',
+  'columns-auto',
+  'break-before-avoid-page',
+  'break-inside-avoid-column',
+  'break-after-column',
+
+  // container
+  //
+
+  // divide
+  'divide',
+  'divide-y-4',
+  'divide-x-4',
+  'divide-x-reverse',
+  'divide-green-500',
+  'divide-opacity-50',
+  'divide-dashed',
+  'divide-dotted',
+  'divide-transparent',
+  'divide-current',
+  'divide-none',
+
+  // filters
+  'filter',
+  'filter-none',
+  'blur',
+  'blur-md',
+  'blur-4',
+  'brightness-0',
+  'brightness-60',
+  'contrast-125',
+  'drop-shadow',
+  'drop-shadow-[0_4px_3px_#000]',
+  'drop-shadow-none',
+  'drop-shadow-md',
+  'grayscale',
+  'grayscale-90',
+  'hue-rotate-0',
+  'hue-rotate-360',
+  'invert',
+  'invert-90',
+  'saturate',
+  'saturate-30',
+  'sepia',
+  'sepia-80',
+
+  // filters - backdrop
+  'backdrop-filter',
+  'backdrop-filter-none',
+  'backdrop-blur',
+  'backdrop-blur-md',
+  'backdrop-blur-4',
+  'backdrop-brightness-0',
+  'backdrop-brightness-60',
+  'backdrop-contrast-125',
+  'backdrop-grayscale',
+  'backdrop-grayscale-90',
+  'backdrop-hue-rotate-0',
+  'backdrop-hue-rotate-360',
+  'backdrop-invert',
+  'backdrop-invert-90',
+  'backdrop-saturate',
+  'backdrop-saturate-30',
+  'backdrop-sepia',
+  'backdrop-sepia-80',
+
+  // line clamp
+  'line-clamp-7',
+  'line-clamp-100',
+  'line-clamp-none',
+
+  // scrolls
+  'snap-y',
+  'snap-mandatory',
+  'snap-none',
+  'snap-center',
+  'snap-align-none',
+  'snap-always',
+  'scroll-m-[3em]',
+  'scroll-m-0',
+  'scroll-m-1/2',
+  'scroll-m-auto',
+  'scroll-p-2',
+  'scroll-p-t-2',
+  'scroll-p2',
+  'scroll-pl-10px',
+  'scroll-pt-2',
+  'scroll-pt2',
+
+  // shadow, spacing
+  'mix-blend-normal',
+  'mix-blend-color-light',
+  'space-x-2',
+  'space-y-4',
+  'space-x-reverse',
+  'space-x-$space',
+
+  // static
+  'capitalize',
+  'normal-case',
+  'hyphens-none',
+  'hyphens-auto',
+  'write-normal',
+  'write-orient-sideways',
+  'not-sr-only',
+  'isolate',
+  'isolate-auto',
+  'object-none',
+  'object-center',
+  'object-ct',
+  'object-cb',
+
+  // tables
+  'border-collapse',
+  'border-separate',
+  'caption-top',
+  'caption-bottom',
+  'inline-table',
+  'table',
+  'table-auto',
+  'table-caption',
+  'table-empty-cells-visible',
+  'table-empty-cells-hidden',
+  'table-footer-group',
+  'table-row-group',
+
+  // touch-actions
+  'touch-pan-left',
+  'touch-pan-y',
+  'touch-pinch-zoom',
+  'touch-manipulation',
+  'touch-none',
+
+  // typography
+  'lining-nums',
+  'normal-nums',
+  'tabular-nums',
+
+  // variables
+  'bg-blend-$data',
+
+  // dark/light
+  '.dark:text-xl',
+  '@dark:text-xl',
+
+  // variants
+  '!animate-ping',
+  '-backdrop-hue-rotate-90',
+  '-hue-rotate-90',
+  'hover:animate-bounce',
+
+  '-scroll-mb-px',
+  '-scroll-p-px',
+  '-space-x-4',
 ]
